support new table modules
